shift - auth listener to conditional router
